pyproject.toml
[project]
name = "skellysubs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arabic-reshaper>=3.0.0",
    "beautifulsoup4>=4.13.3",
    "fastapi[standard]>=0.115.7",
    "ffmpeg-python>=0.2.0",
    "iso639-lang>=2.6.0",
    "jieba>=0.42.1",
    "lxml>=5.3.1",
    "moviepy>=2.1.2",
    "openai>=1.65.1",
    "opencv-python-headless>=4.11.0.86",
    "psutil>=6.1.1",
    "pydantic>=2.10.6",
    "python-bidi>=0.6.3",
    "pytube>=15.0.0",
    "requests>=2.32.3",
    "transformers>=4.48.2",
    "uvicorn>=0.34.0",
    "wikipedia-api>=0.8.1",
    "wikipedia>=1.4.0",
]
